expand(): correct leading zeros issue
Problem: As noted in #32, the expand() function doesn't correctly
handle hostlists with leading zeros in the ranges.

Rework and restructure the expand() function to correctly handle ranges
of leading zeros, particularly when there are multiple ranges.

Add a unit test using the example provided in #32 to confirm it is
working as expected.