cmd: import hostlist module, bump version
Problem: Building the hostlist package with the commands listed in the
start page does not allow for actual use of any of the 'hostlist *'
commands and instead raises an AttributeError complaining that it can't
find any of the functions.

Add the import of the hostlist package to __init__.py so that the
command-line interface can actually utilize the functions defined in
hostlist.py and a user can run things like 'hostlist -e node[1-4]'
instead of 'python cla_hostlist.py -e node[1-4]'.

Bump the version of py-hostlist to 0.1.0.