travis ci: make "pip install" command more general
Problem: The test installation of the hostlist package explicitly
specifies the version number of the hostlist package, but this will no
longer apply when the version number is bumped up.

Change the "pip install" command to instead just use a wildcard so that
we don't have to change the name of the package every time the version
number is bumped.